Relax strict sdl2 version dependency
gfx_window_sdl required sdl2 = "0.29" stoping usage of later sdl2
versions. Since this lib is used as a helper for sdl and isn't
particularly likely to break with newer versions, it is more helpful
to declare a minimum sdl2 version. This allows a root consumer to
declare, for example, sdl2 = "0" and have it work properly.
In the case newer versions cause issues, it is still clear which
version(s) can be tried by viewing the minimum declared.